Simplify GreenNode code layout
